Rename test module to testing to prevent name clash
