+++ ward/__init__.py
@@ -1,3 +1,3 @@
 from .expect import expect, raises
 from .fixtures import fixture
-from .test import skip, xfail
+from .testing import skip, xfail, test
